Move constant filtering policy to providers with shared helper
