+++ core/src/main/java/io/github/lemon_ant/jharmonizer/core/sorter/spoon/dependency_graph/AbstractReferencedFieldsDeclarationDependencyProvider.java
@@ -24,6 +24,9 @@ abstract class AbstractReferencedFieldsDeclarationDependencyProvider implements 
                         DeclaringTypeFieldReferenceUtils.findProviderFieldsRequiredByDependentMember(
                                         dependentMember, ctElement)
                                 .stream()
+                                .filter(providerMember ->
+                                        !CompileTimeConstantFieldUtils.isCompileTimeConstantVariable(
+                                                providerMember))
                                 .map(providerMember -> new MemberDependencyArc(
                                         providerMember, MemberDependencyEdgeKind.DECLARATION_DEPENDENCY))
                                 .collect(Collectors.toUnmodifiableSet()))
