+++ core/src/main/java/io/github/lemon_ant/jharmonizer/core/sorter/spoon/dependency_graph/ExplicitDeclaringTypeInitializerFieldDependencyProvider.java
@@ -18,7 +18,7 @@ final class ExplicitDeclaringTypeInitializerFieldDependencyProvider
     @Override
     protected boolean isSupportedReferencedField(@NonNull CtField<?> referencedField) {
         return isStaticField(referencedField)
-                && !InitializationOrderDependencyUtils.isCompileTimeConstantVariable(referencedField);
+                && !CompileTimeConstantFieldUtils.isCompileTimeConstantVariable(referencedField);
     }
 
     @Override
